Apply single-responsiblity principle
