+++ data/books.json
@@ -6,5 +6,9 @@
   {
     "title": "Takadii",
     "author": "Ticha"
+  },
+  {
+    "title": "To Kill A Mockingbird",
+    "author": "Harper Lee"
   }
 ]
